backend/requirements.txt
alembic==1.16.5
clerk-backend-api==4.2.0
fastapi==0.128.8
httpx==0.28.1
psycopg2-binary==2.9.11
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.2.1
sentry-sdk==2.56.0
SQLAlchemy==2.0.48
slowapi==0.1.9
uvicorn==0.39.0
